Extract ParticipantAvatar to common
