+++ src/com/android/messaging/ui/conversationsettings/common/ConversationSettingsComponents.kt
@@ -1,20 +1,17 @@
 package com.android.messaging.ui.conversationsettings.common
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Block
 import androidx.compose.material.icons.filled.Group
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.Notifications
@@ -27,20 +24,17 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
 import com.android.messaging.R
+import com.android.messaging.ui.common.components.ParticipantAvatar
 import com.android.messaging.ui.conversationsettings.screen.model.ParticipantUiState
 import com.android.messaging.ui.core.AppTheme
 
@@ -57,8 +51,10 @@ internal fun ConversationHeader(
             .padding(bottom = 8.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
+        val isBlocked = participant?.isBlocked == true
+
         ParticipantAvatar(
-            avatarUri = participant?.avatarUri,
+            avatarUri = participant?.avatarUri.takeUnless { isBlocked },
             modifier = Modifier
                 .size(112.dp)
                 .graphicsLayer {
@@ -67,6 +63,7 @@ internal fun ConversationHeader(
                     scaleY = scale
                 },
             fallbackIcon = when {
+                isBlocked -> Icons.Default.Block
                 participant == null -> Icons.Default.Group
                 else -> Icons.Default.Person
             },
@@ -153,9 +150,12 @@ internal fun ParticipantItem(
         verticalAlignment = Alignment.CenterVertically,
     ) {
         ParticipantAvatar(
-            avatarUri = participant.avatarUri,
+            avatarUri = participant.avatarUri.takeUnless { participant.isBlocked },
             modifier = Modifier.size(40.dp),
-            fallbackIcon = Icons.Default.Person,
+            fallbackIcon = when {
+                participant.isBlocked -> Icons.Default.Block
+                else -> Icons.Default.Person
+            },
             fallbackIconSize = 24.dp,
         )
 
@@ -183,41 +183,6 @@ internal fun ParticipantItem(
                     imageVector = Icons.Default.Info,
                     contentDescription = stringResource(R.string.action_contact_info),
                     tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-internal fun ParticipantAvatar(
-    avatarUri: String?,
-    fallbackIcon: ImageVector,
-    fallbackIconSize: Dp,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier
-            .clip(CircleShape)
-            .background(MaterialTheme.colorScheme.primaryContainer),
-        contentAlignment = Alignment.Center,
-    ) {
-        when {
-            avatarUri.isNullOrBlank() -> {
-                Icon(
-                    imageVector = fallbackIcon,
-                    contentDescription = null,
-                    modifier = Modifier.size(fallbackIconSize),
-                    tint = MaterialTheme.colorScheme.onPrimaryContainer,
-                )
-            }
-
-            else -> {
-                AsyncImage(
-                    model = avatarUri,
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier.fillMaxSize(),
                 )
             }
         }
