+++ src/com/android/messaging/ui/conversationsettings/screen/model/ConversationSettingsNavEvent.kt
@@ -7,6 +7,4 @@ internal sealed interface ConversationSettingsNavEvent {
     ) : ConversationSettingsNavEvent
 
     data object CloseAfterArchive : ConversationSettingsNavEvent
-
-    data object CloseAfterBlock : ConversationSettingsNavEvent
 }
